+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-05-15 12:11:23
-// Commit: fix(core): correct layout constraints and overflow in core
+// Updated at: 2026-05-16 20:05:19
+// Commit: feat(core): integrate custom telemetry logger in core
 
 class coreComponentState {
-  final String commitRef = '1179';
-  final String lastMessage = 'fix(core): correct layout constraints and overflow in core';
-  final DateTime timestamp = DateTime.parse('2026-05-15T12:11:23+05:00');
+  final String commitRef = '1211';
+  final String lastMessage = 'feat(core): integrate custom telemetry logger in core';
+  final DateTime timestamp = DateTime.parse('2026-05-16T20:05:19+05:00');
 }